fix(po,cli): reject negative input for unsigned numeric options
`--gas-limit` and `--memory-page-limit` are declared as `PO::List<int>`
in `include/driver/tool.h`, but consumed as `uint64_t` via
`setCostLimit` and `setMaxMemoryPage` after a lossy
`static_cast<uint32_t>`. Two user-visible bugs follow:

  * `--gas-limit=-1` is silently accepted; the cast yields 4294967295
    instead of rejecting the invalid input.
  * `--gas-limit=9999999999` (10 billion, well within `uint64_t`) is
    rejected with "out of range" because the parser is targeting
    `int`, capping the user at ~2.1 billion.

`TimeLim` next door is already `PO::Option<uint64_t>`; `GasLim` and
`MemLim` were the outliers.

Switching the field types to `PO::List<uint64_t>` and dropping the
`static_cast<uint32_t>` truncation is not enough on its own:
`strtoul`/`strtoull` accept a leading minus and silently wrap it to
`ULLONG_MAX` without setting `ERANGE`, so `--gas-limit=-1` would still
be accepted (just as "unlimited" instead of 4G). Harden
`PO::stringToInteger` to reject a leading `-` when the result type is
unsigned, so every current and future unsigned option benefits from the
check.

After this change:

  * `--gas-limit=-1` -> "integer value out of range: -1"
  * `--gas-limit=18446744073709551615` -> accepted (full uint64_t range)
  * `--gas-limit=18446744073709551616` -> "integer value out of range"

Add parametric coverage in `test/po/po.cpp` for `List<uint64_t>` to
pin down the new behaviour: full-range positives accepted, negatives
rejected, overflow rejected, non-numeric rejected.

+++ include/driver/tool.h
@@ -183,8 +183,8 @@ struct DriverToolOptions {
   PO::Option<std::string> ConfRunMode;
   PO::Option<PO::Toggle> ConfAFUNIX;
   PO::Option<uint64_t> TimeLim;
-  PO::List<int> GasLim;
-  PO::List<int> MemLim;
+  PO::List<uint64_t> GasLim;
+  PO::List<uint64_t> MemLim;
   PO::List<std::string> ForbiddenPlugins;
   PO::Option<std::string> LogLevel;
