+++ lib/driver/fuzzPO.cpp
@@ -194,12 +194,12 @@ int FuzzPO(const uint8_t *Data, size_t Size) noexcept {
           "Limitation of maximum time(in milliseconds) for execution, default value is 0 for no limitations"sv),
       PO::MetaVar("TIMEOUT"sv), PO::DefaultValue<uint64_t>(0));
 
-  PO::List<int> GasLim(
+  PO::List<uint64_t> GasLim(
       PO::Description(
           "Limitation of execution gas. Upper bound can be specified as --gas-limit `GAS_LIMIT`."sv),
       PO::MetaVar("GAS_LIMIT"sv));
 
-  PO::List<int> MemLim(
+  PO::List<uint64_t> MemLim(
       PO::Description(
           "Limitation of pages(as size of 64 KiB) in every memory instance. Upper bound can be specified as --memory-page-limit `PAGE_COUNT`."sv),
       PO::MetaVar("PAGE_COUNT"sv));
